Move AgentConfig to dedicated file
Trying to reduce some complexity from the Agent
